+++ src/lib/components/Loading.tsx
@@ -1,10 +1,10 @@
-import { Spinner, Text } from "@chakra-ui/react";
+import { Flex, Spinner, Text } from "@chakra-ui/react";
 
 export const Loading = () => {
   return (
-    <>
+    <Flex borderY="1px solid" borderColor="divider.main" width="full" py="48px">
       <Spinner size="xl" speed="0.65s" />
       <Text mt="20px">Loading ...</Text>
-    </>
+    </Flex>
   );
 };